Ensure we have an orgId in withAdminAccess function or 500
In the future this won't happen, as we'll properly set the orgid cookie
on signup.

A better future fix is to store the org id in an encrypted cookie or the
session token cookie. This was previously not feasible bc the frontend
used to set the org id cookie directly, but this may not be a blocker anymore.